+++ src/main/java/cz/sparko/boxitory/controller/GlobalControllerExceptionHandler.java
@@ -8,14 +8,13 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 
 @ControllerAdvice
 public class GlobalControllerExceptionHandler {
-    @ExceptionHandler
+    @ExceptionHandler(NotFoundException.class)
+    public ResponseEntity<String> handleException(NotFoundException e) {
+        return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
+    }
+
+    @ExceptionHandler(Exception.class)
     public ResponseEntity<String> handleException(Exception e) {
-        final HttpStatus status;
-        if (e instanceof NotFoundException) {
-            status = HttpStatus.NOT_FOUND;
-        } else {
-            status = HttpStatus.INTERNAL_SERVER_ERROR;
-        }
-        return new ResponseEntity<>(e.getMessage(), status);
+        return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
     }
 }
