+++ src/test/java/cz/sparko/boxitory/test/e2e/AbstractIntegrationTest.java
@@ -3,6 +3,7 @@ package cz.sparko.boxitory.test.e2e;
 import cz.sparko.boxitory.App;
 import cz.sparko.boxitory.conf.AppProperties;
 import cz.sparko.boxitory.controller.BoxController;
+import cz.sparko.boxitory.controller.BoxRestController;
 import cz.sparko.boxitory.controller.DownloadController;
 import org.apache.commons.io.FileUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,7 +23,7 @@ import java.io.File;
 import java.io.IOException;
 
 @ContextConfiguration(classes = App.class)
-@WebMvcTest(controllers = {BoxController.class, DownloadController.class})
+@WebMvcTest(controllers = {BoxRestController.class, DownloadController.class, BoxController.class})
 @AutoConfigureMockMvc(secure = false)
 @TestPropertySource(locations = "classpath:test.properties")
 @Test(groups = "e2e")
